qrtrex/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.3
daphne==4.2.0
Django==5.2.1
django-pwa==2.0.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pywin32==310
qrcode==8.2
razorpay==1.4.2
redis==6.1.0
reportlab==4.4.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.4.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
zope.interface==7.2
